requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-pinecone==0.2.6
pinecone==6.0.0
ctransformers==0.2.27
sentence-transformers
flask
pypdf
python-dotenv


-e .